All refuge restrooms

All listed refuge restrooms, ordered by date added or last updated, with the most recent additions and updates at the top.

rfg_all_restrooms(accessible = FALSE, unisex = FALSE, verbose = TRUE,
  tidy = FALSE)

Arguments

accessible
If TRUE, only returns restrooms that meet the Americans with Disability Act standards for accessibility. Defaults to FALSE.

unisex
If TRUE, only returns unisex restrooms. Defaults to FALSE.

verbose
If TRUE, prints query progress. Defaults to TRUE.

tidy
If TRUE, makes USA state names more consistent. The internal function focuses on correcting known errors and mistakes, if incorrect state data is supplied that cannot be corrected. Defaults to FALSE.

Value

A tibble with details on all listed bathrooms.

Details

Note that the API does not standardise or validate the names of states, cities, countries, etc, and so this data may need to be cleaned and organised.

Examples

a <- rfg_all_restrooms(accessible = TRUE, unisex = TRUE)

#> Retrieving page 1 of 176

#> Retrieving page 2 of 176

#> Retrieving page 3 of 176

#> Retrieving page 4 of 176

#> Retrieving page 5 of 176

#> Retrieving page 6 of 176

#> Retrieving page 7 of 176

#> Retrieving page 8 of 176

#> Retrieving page 9 of 176

#> Retrieving page 10 of 176

#> Retrieving page 11 of 176

#> Retrieving page 12 of 176

#> Retrieving page 13 of 176

#> Retrieving page 14 of 176

#> Retrieving page 15 of 176

#> Retrieving page 16 of 176

#> Retrieving page 17 of 176

#> Retrieving page 18 of 176

#> Retrieving page 19 of 176

#> Retrieving page 20 of 176

#> Retrieving page 21 of 176

#> Retrieving page 22 of 176

#> Retrieving page 23 of 176

#> Retrieving page 24 of 176

#> Retrieving page 25 of 176

#> Retrieving page 26 of 176

#> Retrieving page 27 of 176

#> Retrieving page 28 of 176

#> Retrieving page 29 of 176

#> Retrieving page 30 of 176

#> Retrieving page 31 of 176

#> Retrieving page 32 of 176

#> Retrieving page 33 of 176

#> Retrieving page 34 of 176

#> Retrieving page 35 of 176

#> Retrieving page 36 of 176

#> Retrieving page 37 of 176

#> Retrieving page 38 of 176

#> Retrieving page 39 of 176

#> Retrieving page 40 of 176

#> Retrieving page 41 of 176

#> Retrieving page 42 of 176

#> Retrieving page 43 of 176

#> Retrieving page 44 of 176

#> Retrieving page 45 of 176

#> Retrieving page 46 of 176

#> Retrieving page 47 of 176

#> Retrieving page 48 of 176

#> Retrieving page 49 of 176

#> Retrieving page 50 of 176

#> Retrieving page 51 of 176

#> Retrieving page 52 of 176

#> Retrieving page 53 of 176

#> Retrieving page 54 of 176

#> Retrieving page 55 of 176

#> Retrieving page 56 of 176

#> Retrieving page 57 of 176

#> Retrieving page 58 of 176

#> Retrieving page 59 of 176

#> Retrieving page 60 of 176

#> Retrieving page 61 of 176

#> Retrieving page 62 of 176

#> Retrieving page 63 of 176

#> Retrieving page 64 of 176

#> Retrieving page 65 of 176

#> Retrieving page 66 of 176

#> Retrieving page 67 of 176

#> Retrieving page 68 of 176

#> Retrieving page 69 of 176

#> Retrieving page 70 of 176

#> Retrieving page 71 of 176

#> Retrieving page 72 of 176

#> Retrieving page 73 of 176

#> Retrieving page 74 of 176

#> Retrieving page 75 of 176

#> Retrieving page 76 of 176

#> Retrieving page 77 of 176

#> Retrieving page 78 of 176

#> Retrieving page 79 of 176

#> Retrieving page 80 of 176

#> Retrieving page 81 of 176

#> Retrieving page 82 of 176

#> Retrieving page 83 of 176

#> Retrieving page 84 of 176

#> Retrieving page 85 of 176

#> Retrieving page 86 of 176

#> Retrieving page 87 of 176

#> Retrieving page 88 of 176

#> Retrieving page 89 of 176

#> Retrieving page 90 of 176

#> Retrieving page 91 of 176

#> Retrieving page 92 of 176

#> Retrieving page 93 of 176

#> Retrieving page 94 of 176

#> Retrieving page 95 of 176

#> Retrieving page 96 of 176

#> Retrieving page 97 of 176

#> Retrieving page 98 of 176

#> Retrieving page 99 of 176

#> Retrieving page 100 of 176

#> Retrieving page 101 of 176

#> Retrieving page 102 of 176

#> Retrieving page 103 of 176

#> Retrieving page 104 of 176

#> Retrieving page 105 of 176

#> Retrieving page 106 of 176

#> Retrieving page 107 of 176

#> Retrieving page 108 of 176

#> Retrieving page 109 of 176

#> Retrieving page 110 of 176

#> Retrieving page 111 of 176

#> Retrieving page 112 of 176

#> Retrieving page 113 of 176

#> Retrieving page 114 of 176

#> Retrieving page 115 of 176

#> Retrieving page 116 of 176

#> Retrieving page 117 of 176

#> Retrieving page 118 of 176

#> Retrieving page 119 of 176

#> Retrieving page 120 of 176

#> Retrieving page 121 of 176

#> Retrieving page 122 of 176

#> Retrieving page 123 of 176

#> Retrieving page 124 of 176

#> Retrieving page 125 of 176

#> Retrieving page 126 of 176

#> Retrieving page 127 of 176

#> Retrieving page 128 of 176

#> Retrieving page 129 of 176

#> Retrieving page 130 of 176

#> Retrieving page 131 of 176

#> Retrieving page 132 of 176

#> Retrieving page 133 of 176

#> Retrieving page 134 of 176

#> Retrieving page 135 of 176

#> Retrieving page 136 of 176

#> Retrieving page 137 of 176

#> Retrieving page 138 of 176

#> Retrieving page 139 of 176

#> Retrieving page 140 of 176

#> Retrieving page 141 of 176

#> Retrieving page 142 of 176

#> Retrieving page 143 of 176

#> Retrieving page 144 of 176

#> Retrieving page 145 of 176

#> Retrieving page 146 of 176

#> Retrieving page 147 of 176

#> Retrieving page 148 of 176

#> Retrieving page 149 of 176

#> Retrieving page 150 of 176

#> Retrieving page 151 of 176

#> Retrieving page 152 of 176

#> Retrieving page 153 of 176

#> Retrieving page 154 of 176

#> Retrieving page 155 of 176

#> Retrieving page 156 of 176

#> Retrieving page 157 of 176

#> Retrieving page 158 of 176

#> Retrieving page 159 of 176

#> Retrieving page 160 of 176

#> Retrieving page 161 of 176

#> Retrieving page 162 of 176

#> Retrieving page 163 of 176

#> Retrieving page 164 of 176

#> Retrieving page 165 of 176

#> Retrieving page 166 of 176

#> Retrieving page 167 of 176

#> Retrieving page 168 of 176

#> Retrieving page 169 of 176

#> Retrieving page 170 of 176

#> Retrieving page 171 of 176

#> Retrieving page 172 of 176

#> Retrieving page 173 of 176

#> Retrieving page 174 of 176

#> Retrieving page 175 of 176

#> Retrieving page 176 of 176