Bathrooms by dates

Requests all records on bathrooms created or updated on or after a given date.

rfg_date(date = NULL, accessible = FALSE, unisex = FALSE,
  verbose = TRUE)

Arguments

date
A date in "yyyy-mm-dd" format, or any format that can be coerced to a date with as.Date().

accessible
If TRUE, only returns restrooms that meet the Americans with Disability Act standards for accessibility. Defaults to FALSE.

unisex
If TRUE, only returns unisex restrooms. Defaults to FALSE.

verbose
If TRUE, prints query progress. Defaults to TRUE.

Value

A tibble with all bathrooms recorded or updated on or after the given date.

Examples

q <- rfg_date("2017-11-04")

#> Retrieving page 1 of 308

#> Retrieving page 2 of 308

#> Retrieving page 3 of 308

#> Retrieving page 4 of 308

#> Retrieving page 5 of 308

#> Retrieving page 6 of 308

#> Retrieving page 7 of 308

#> Retrieving page 8 of 308

#> Retrieving page 9 of 308

#> Retrieving page 10 of 308

#> Retrieving page 11 of 308

#> Retrieving page 12 of 308

#> Retrieving page 13 of 308

#> Retrieving page 14 of 308

#> Retrieving page 15 of 308

#> Retrieving page 16 of 308

#> Retrieving page 17 of 308

#> Retrieving page 18 of 308

#> Retrieving page 19 of 308

#> Retrieving page 20 of 308

#> Retrieving page 21 of 308

#> Retrieving page 22 of 308

#> Retrieving page 23 of 308

#> Retrieving page 24 of 308

#> Retrieving page 25 of 308

#> Retrieving page 26 of 308

#> Retrieving page 27 of 308

#> Retrieving page 28 of 308

#> Retrieving page 29 of 308

#> Retrieving page 30 of 308

#> Retrieving page 31 of 308

#> Retrieving page 32 of 308

#> Retrieving page 33 of 308

#> Retrieving page 34 of 308

#> Retrieving page 35 of 308

#> Retrieving page 36 of 308

#> Retrieving page 37 of 308

#> Retrieving page 38 of 308

#> Retrieving page 39 of 308

#> Retrieving page 40 of 308

#> Retrieving page 41 of 308

#> Retrieving page 42 of 308

#> Retrieving page 43 of 308

#> Retrieving page 44 of 308

#> Retrieving page 45 of 308

#> Retrieving page 46 of 308

#> Retrieving page 47 of 308

#> Retrieving page 48 of 308

#> Retrieving page 49 of 308

#> Retrieving page 50 of 308

#> Retrieving page 51 of 308

#> Retrieving page 52 of 308

#> Retrieving page 53 of 308

#> Retrieving page 54 of 308

#> Retrieving page 55 of 308

#> Retrieving page 56 of 308

#> Retrieving page 57 of 308

#> Retrieving page 58 of 308

#> Retrieving page 59 of 308

#> Retrieving page 60 of 308

#> Retrieving page 61 of 308

#> Retrieving page 62 of 308

#> Retrieving page 63 of 308

#> Retrieving page 64 of 308

#> Retrieving page 65 of 308

#> Retrieving page 66 of 308

#> Retrieving page 67 of 308

#> Retrieving page 68 of 308

#> Retrieving page 69 of 308

#> Retrieving page 70 of 308

#> Retrieving page 71 of 308

#> Retrieving page 72 of 308

#> Retrieving page 73 of 308

#> Retrieving page 74 of 308

#> Retrieving page 75 of 308

#> Retrieving page 76 of 308

#> Retrieving page 77 of 308

#> Retrieving page 78 of 308

#> Retrieving page 79 of 308

#> Retrieving page 80 of 308

#> Retrieving page 81 of 308

#> Retrieving page 82 of 308

#> Retrieving page 83 of 308

#> Retrieving page 84 of 308

#> Retrieving page 85 of 308

#> Retrieving page 86 of 308

#> Retrieving page 87 of 308

#> Retrieving page 88 of 308

#> Retrieving page 89 of 308

#> Retrieving page 90 of 308

#> Retrieving page 91 of 308

#> Retrieving page 92 of 308

#> Retrieving page 93 of 308

#> Retrieving page 94 of 308

#> Retrieving page 95 of 308

#> Retrieving page 96 of 308

#> Retrieving page 97 of 308

#> Retrieving page 98 of 308

#> Retrieving page 99 of 308

#> Retrieving page 100 of 308

#> Retrieving page 101 of 308

#> Retrieving page 102 of 308

#> Retrieving page 103 of 308

#> Retrieving page 104 of 308

#> Retrieving page 105 of 308

#> Retrieving page 106 of 308

#> Retrieving page 107 of 308

#> Retrieving page 108 of 308

#> Retrieving page 109 of 308

#> Retrieving page 110 of 308

#> Retrieving page 111 of 308

#> Retrieving page 112 of 308

#> Retrieving page 113 of 308

#> Retrieving page 114 of 308

#> Retrieving page 115 of 308

#> Retrieving page 116 of 308

#> Retrieving page 117 of 308

#> Retrieving page 118 of 308

#> Retrieving page 119 of 308

#> Retrieving page 120 of 308

#> Retrieving page 121 of 308

#> Retrieving page 122 of 308

#> Retrieving page 123 of 308

#> Retrieving page 124 of 308

#> Retrieving page 125 of 308

#> Retrieving page 126 of 308

#> Retrieving page 127 of 308

#> Retrieving page 128 of 308

#> Retrieving page 129 of 308

#> Retrieving page 130 of 308

#> Retrieving page 131 of 308

#> Retrieving page 132 of 308

#> Retrieving page 133 of 308

#> Retrieving page 134 of 308

#> Retrieving page 135 of 308

#> Retrieving page 136 of 308

#> Retrieving page 137 of 308

#> Retrieving page 138 of 308

#> Retrieving page 139 of 308

#> Retrieving page 140 of 308

#> Retrieving page 141 of 308

#> Retrieving page 142 of 308

#> Retrieving page 143 of 308

#> Retrieving page 144 of 308

#> Retrieving page 145 of 308

#> Retrieving page 146 of 308

#> Retrieving page 147 of 308

#> Retrieving page 148 of 308

#> Retrieving page 149 of 308

#> Retrieving page 150 of 308

#> Retrieving page 151 of 308

#> Retrieving page 152 of 308

#> Retrieving page 153 of 308

#> Retrieving page 154 of 308

#> Retrieving page 155 of 308

#> Retrieving page 156 of 308

#> Retrieving page 157 of 308

#> Retrieving page 158 of 308

#> Retrieving page 159 of 308

#> Retrieving page 160 of 308

#> Retrieving page 161 of 308

#> Retrieving page 162 of 308

#> Retrieving page 163 of 308

#> Retrieving page 164 of 308

#> Retrieving page 165 of 308

#> Retrieving page 166 of 308

#> Retrieving page 167 of 308

#> Retrieving page 168 of 308

#> Retrieving page 169 of 308

#> Retrieving page 170 of 308

#> Retrieving page 171 of 308

#> Retrieving page 172 of 308

#> Retrieving page 173 of 308

#> Retrieving page 174 of 308

#> Retrieving page 175 of 308

#> Retrieving page 176 of 308

#> Retrieving page 177 of 308

#> Retrieving page 178 of 308

#> Retrieving page 179 of 308

#> Retrieving page 180 of 308

#> Retrieving page 181 of 308

#> Retrieving page 182 of 308

#> Retrieving page 183 of 308

#> Retrieving page 184 of 308

#> Retrieving page 185 of 308

#> Retrieving page 186 of 308

#> Retrieving page 187 of 308

#> Retrieving page 188 of 308

#> Retrieving page 189 of 308

#> Retrieving page 190 of 308

#> Retrieving page 191 of 308

#> Retrieving page 192 of 308

#> Retrieving page 193 of 308

#> Retrieving page 194 of 308

#> Retrieving page 195 of 308

#> Retrieving page 196 of 308

#> Retrieving page 197 of 308

#> Retrieving page 198 of 308

#> Retrieving page 199 of 308

#> Retrieving page 200 of 308

#> Retrieving page 201 of 308

#> Retrieving page 202 of 308

#> Retrieving page 203 of 308

#> Retrieving page 204 of 308

#> Retrieving page 205 of 308

#> Retrieving page 206 of 308

#> Retrieving page 207 of 308

#> Retrieving page 208 of 308

#> Retrieving page 209 of 308

#> Retrieving page 210 of 308

#> Retrieving page 211 of 308

#> Retrieving page 212 of 308

#> Retrieving page 213 of 308

#> Retrieving page 214 of 308

#> Retrieving page 215 of 308

#> Retrieving page 216 of 308

#> Retrieving page 217 of 308

#> Retrieving page 218 of 308

#> Retrieving page 219 of 308

#> Retrieving page 220 of 308

#> Retrieving page 221 of 308

#> Retrieving page 222 of 308

#> Retrieving page 223 of 308

#> Retrieving page 224 of 308

#> Retrieving page 225 of 308

#> Retrieving page 226 of 308

#> Retrieving page 227 of 308

#> Retrieving page 228 of 308

#> Retrieving page 229 of 308

#> Retrieving page 230 of 308

#> Retrieving page 231 of 308

#> Retrieving page 232 of 308

#> Retrieving page 233 of 308

#> Retrieving page 234 of 308

#> Retrieving page 235 of 308

#> Retrieving page 236 of 308

#> Retrieving page 237 of 308

#> Retrieving page 238 of 308

#> Retrieving page 239 of 308

#> Retrieving page 240 of 308

#> Retrieving page 241 of 308

#> Retrieving page 242 of 308

#> Retrieving page 243 of 308

#> Retrieving page 244 of 308

#> Retrieving page 245 of 308

#> Retrieving page 246 of 308

#> Retrieving page 247 of 308

#> Retrieving page 248 of 308

#> Retrieving page 249 of 308

#> Retrieving page 250 of 308

#> Retrieving page 251 of 308

#> Retrieving page 252 of 308

#> Retrieving page 253 of 308

#> Retrieving page 254 of 308

#> Retrieving page 255 of 308

#> Retrieving page 256 of 308

#> Retrieving page 257 of 308

#> Retrieving page 258 of 308

#> Retrieving page 259 of 308

#> Retrieving page 260 of 308

#> Retrieving page 261 of 308

#> Retrieving page 262 of 308

#> Retrieving page 263 of 308

#> Retrieving page 264 of 308

#> Retrieving page 265 of 308

#> Retrieving page 266 of 308

#> Retrieving page 267 of 308

#> Retrieving page 268 of 308

#> Retrieving page 269 of 308

#> Retrieving page 270 of 308

#> Retrieving page 271 of 308

#> Retrieving page 272 of 308

#> Retrieving page 273 of 308

#> Retrieving page 274 of 308

#> Retrieving page 275 of 308

#> Retrieving page 276 of 308

#> Retrieving page 277 of 308

#> Retrieving page 278 of 308

#> Retrieving page 279 of 308

#> Retrieving page 280 of 308

#> Retrieving page 281 of 308

#> Retrieving page 282 of 308

#> Retrieving page 283 of 308

#> Retrieving page 284 of 308

#> Retrieving page 285 of 308

#> Retrieving page 286 of 308

#> Retrieving page 287 of 308

#> Retrieving page 288 of 308

#> Retrieving page 289 of 308

#> Retrieving page 290 of 308

#> Retrieving page 291 of 308

#> Retrieving page 292 of 308

#> Retrieving page 293 of 308

#> Retrieving page 294 of 308

#> Retrieving page 295 of 308

#> Retrieving page 296 of 308

#> Retrieving page 297 of 308

#> Retrieving page 298 of 308

#> Retrieving page 299 of 308

#> Retrieving page 300 of 308

#> Retrieving page 301 of 308

#> Retrieving page 302 of 308

#> Retrieving page 303 of 308

#> Retrieving page 304 of 308

#> Retrieving page 305 of 308

#> Retrieving page 306 of 308

#> Retrieving page 307 of 308

#> Retrieving page 308 of 308