Bathrooms by dates

Requests all records on bathrooms created or updated on or after a given date.

rfg_date(date = NULL, accessible = FALSE, unisex = FALSE,
  verbose = TRUE, tidy = FALSE)

Arguments

date
A date in "yyyy-mm-dd" format, or any format that can be coerced to a date with as.Date().

accessible
If TRUE, only returns restrooms that meet the Americans with Disability Act standards for accessibility. Defaults to FALSE.

unisex
If TRUE, only returns unisex restrooms. Defaults to FALSE.

verbose
If TRUE, prints query progress. Defaults to TRUE.

tidy
If TRUE, makes USA state names more consistent. The internal function focuses on correcting known errors and mistakes, if incorrect state data is supplied that cannot be corrected. Defaults to FALSE.

Value

A tibble with all bathrooms recorded or updated on or after the given date.

Examples

q <- rfg_date("2017-11-04")

#> Retrieving page 1 of 349

#> Retrieving page 2 of 349

#> Retrieving page 3 of 349

#> Retrieving page 4 of 349

#> Retrieving page 5 of 349

#> Retrieving page 6 of 349

#> Retrieving page 7 of 349

#> Retrieving page 8 of 349

#> Retrieving page 9 of 349

#> Retrieving page 10 of 349

#> Retrieving page 11 of 349

#> Retrieving page 12 of 349

#> Retrieving page 13 of 349

#> Retrieving page 14 of 349

#> Retrieving page 15 of 349

#> Retrieving page 16 of 349

#> Retrieving page 17 of 349

#> Retrieving page 18 of 349

#> Retrieving page 19 of 349

#> Retrieving page 20 of 349

#> Retrieving page 21 of 349

#> Retrieving page 22 of 349

#> Retrieving page 23 of 349

#> Retrieving page 24 of 349

#> Retrieving page 25 of 349

#> Retrieving page 26 of 349

#> Retrieving page 27 of 349

#> Retrieving page 28 of 349

#> Retrieving page 29 of 349

#> Retrieving page 30 of 349

#> Retrieving page 31 of 349

#> Retrieving page 32 of 349

#> Retrieving page 33 of 349

#> Retrieving page 34 of 349

#> Retrieving page 35 of 349

#> Retrieving page 36 of 349

#> Retrieving page 37 of 349

#> Retrieving page 38 of 349

#> Retrieving page 39 of 349

#> Retrieving page 40 of 349

#> Retrieving page 41 of 349

#> Retrieving page 42 of 349

#> Retrieving page 43 of 349

#> Retrieving page 44 of 349

#> Retrieving page 45 of 349

#> Retrieving page 46 of 349

#> Retrieving page 47 of 349

#> Retrieving page 48 of 349

#> Retrieving page 49 of 349

#> Retrieving page 50 of 349

#> Retrieving page 51 of 349

#> Retrieving page 52 of 349

#> Retrieving page 53 of 349

#> Retrieving page 54 of 349

#> Retrieving page 55 of 349

#> Retrieving page 56 of 349

#> Retrieving page 57 of 349

#> Retrieving page 58 of 349

#> Retrieving page 59 of 349

#> Retrieving page 60 of 349

#> Retrieving page 61 of 349

#> Retrieving page 62 of 349

#> Retrieving page 63 of 349

#> Retrieving page 64 of 349

#> Retrieving page 65 of 349

#> Retrieving page 66 of 349

#> Retrieving page 67 of 349

#> Retrieving page 68 of 349

#> Retrieving page 69 of 349

#> Retrieving page 70 of 349

#> Retrieving page 71 of 349

#> Retrieving page 72 of 349

#> Retrieving page 73 of 349

#> Retrieving page 74 of 349

#> Retrieving page 75 of 349

#> Retrieving page 76 of 349

#> Retrieving page 77 of 349

#> Retrieving page 78 of 349

#> Retrieving page 79 of 349

#> Retrieving page 80 of 349

#> Retrieving page 81 of 349

#> Retrieving page 82 of 349

#> Retrieving page 83 of 349

#> Retrieving page 84 of 349

#> Retrieving page 85 of 349

#> Retrieving page 86 of 349

#> Retrieving page 87 of 349

#> Retrieving page 88 of 349

#> Retrieving page 89 of 349

#> Retrieving page 90 of 349

#> Retrieving page 91 of 349

#> Retrieving page 92 of 349

#> Retrieving page 93 of 349

#> Retrieving page 94 of 349

#> Retrieving page 95 of 349

#> Retrieving page 96 of 349

#> Retrieving page 97 of 349

#> Retrieving page 98 of 349

#> Retrieving page 99 of 349

#> Retrieving page 100 of 349

#> Retrieving page 101 of 349

#> Retrieving page 102 of 349

#> Retrieving page 103 of 349

#> Retrieving page 104 of 349

#> Retrieving page 105 of 349

#> Retrieving page 106 of 349

#> Retrieving page 107 of 349

#> Retrieving page 108 of 349

#> Retrieving page 109 of 349

#> Retrieving page 110 of 349

#> Retrieving page 111 of 349

#> Retrieving page 112 of 349

#> Retrieving page 113 of 349

#> Retrieving page 114 of 349

#> Retrieving page 115 of 349

#> Retrieving page 116 of 349

#> Retrieving page 117 of 349

#> Retrieving page 118 of 349

#> Retrieving page 119 of 349

#> Retrieving page 120 of 349

#> Retrieving page 121 of 349

#> Retrieving page 122 of 349

#> Retrieving page 123 of 349

#> Retrieving page 124 of 349

#> Retrieving page 125 of 349

#> Retrieving page 126 of 349

#> Retrieving page 127 of 349

#> Retrieving page 128 of 349

#> Retrieving page 129 of 349

#> Retrieving page 130 of 349

#> Retrieving page 131 of 349

#> Retrieving page 132 of 349

#> Retrieving page 133 of 349

#> Retrieving page 134 of 349

#> Retrieving page 135 of 349

#> Retrieving page 136 of 349

#> Retrieving page 137 of 349

#> Retrieving page 138 of 349

#> Retrieving page 139 of 349

#> Retrieving page 140 of 349

#> Retrieving page 141 of 349

#> Retrieving page 142 of 349

#> Retrieving page 143 of 349

#> Retrieving page 144 of 349

#> Retrieving page 145 of 349

#> Retrieving page 146 of 349

#> Retrieving page 147 of 349

#> Retrieving page 148 of 349

#> Retrieving page 149 of 349

#> Retrieving page 150 of 349

#> Retrieving page 151 of 349

#> Retrieving page 152 of 349

#> Retrieving page 153 of 349

#> Retrieving page 154 of 349

#> Retrieving page 155 of 349

#> Retrieving page 156 of 349

#> Retrieving page 157 of 349

#> Retrieving page 158 of 349

#> Retrieving page 159 of 349

#> Retrieving page 160 of 349

#> Retrieving page 161 of 349

#> Retrieving page 162 of 349

#> Retrieving page 163 of 349

#> Retrieving page 164 of 349

#> Retrieving page 165 of 349

#> Retrieving page 166 of 349

#> Retrieving page 167 of 349

#> Retrieving page 168 of 349

#> Retrieving page 169 of 349

#> Retrieving page 170 of 349

#> Retrieving page 171 of 349

#> Retrieving page 172 of 349

#> Retrieving page 173 of 349

#> Retrieving page 174 of 349

#> Retrieving page 175 of 349

#> Retrieving page 176 of 349

#> Retrieving page 177 of 349

#> Retrieving page 178 of 349

#> Retrieving page 179 of 349

#> Retrieving page 180 of 349

#> Retrieving page 181 of 349

#> Retrieving page 182 of 349

#> Retrieving page 183 of 349

#> Retrieving page 184 of 349

#> Retrieving page 185 of 349

#> Retrieving page 186 of 349

#> Retrieving page 187 of 349

#> Retrieving page 188 of 349

#> Retrieving page 189 of 349

#> Retrieving page 190 of 349

#> Retrieving page 191 of 349

#> Retrieving page 192 of 349

#> Retrieving page 193 of 349

#> Retrieving page 194 of 349

#> Retrieving page 195 of 349

#> Retrieving page 196 of 349

#> Retrieving page 197 of 349

#> Retrieving page 198 of 349

#> Retrieving page 199 of 349

#> Retrieving page 200 of 349

#> Retrieving page 201 of 349

#> Retrieving page 202 of 349

#> Retrieving page 203 of 349

#> Retrieving page 204 of 349

#> Retrieving page 205 of 349

#> Retrieving page 206 of 349

#> Retrieving page 207 of 349

#> Retrieving page 208 of 349

#> Retrieving page 209 of 349

#> Retrieving page 210 of 349

#> Retrieving page 211 of 349

#> Retrieving page 212 of 349

#> Retrieving page 213 of 349

#> Retrieving page 214 of 349

#> Retrieving page 215 of 349

#> Retrieving page 216 of 349

#> Retrieving page 217 of 349

#> Retrieving page 218 of 349

#> Retrieving page 219 of 349

#> Retrieving page 220 of 349

#> Retrieving page 221 of 349

#> Retrieving page 222 of 349

#> Retrieving page 223 of 349

#> Retrieving page 224 of 349

#> Retrieving page 225 of 349

#> Retrieving page 226 of 349

#> Retrieving page 227 of 349

#> Retrieving page 228 of 349

#> Retrieving page 229 of 349

#> Retrieving page 230 of 349

#> Retrieving page 231 of 349

#> Retrieving page 232 of 349

#> Retrieving page 233 of 349

#> Retrieving page 234 of 349

#> Retrieving page 235 of 349

#> Retrieving page 236 of 349

#> Retrieving page 237 of 349

#> Retrieving page 238 of 349

#> Retrieving page 239 of 349

#> Retrieving page 240 of 349

#> Retrieving page 241 of 349

#> Retrieving page 242 of 349

#> Retrieving page 243 of 349

#> Retrieving page 244 of 349

#> Retrieving page 245 of 349

#> Retrieving page 246 of 349

#> Retrieving page 247 of 349

#> Retrieving page 248 of 349

#> Retrieving page 249 of 349

#> Retrieving page 250 of 349

#> Retrieving page 251 of 349

#> Retrieving page 252 of 349

#> Retrieving page 253 of 349

#> Retrieving page 254 of 349

#> Retrieving page 255 of 349

#> Retrieving page 256 of 349

#> Retrieving page 257 of 349

#> Retrieving page 258 of 349

#> Retrieving page 259 of 349

#> Retrieving page 260 of 349

#> Retrieving page 261 of 349

#> Retrieving page 262 of 349

#> Retrieving page 263 of 349

#> Retrieving page 264 of 349

#> Retrieving page 265 of 349

#> Retrieving page 266 of 349

#> Retrieving page 267 of 349

#> Retrieving page 268 of 349

#> Retrieving page 269 of 349

#> Retrieving page 270 of 349

#> Retrieving page 271 of 349

#> Retrieving page 272 of 349

#> Retrieving page 273 of 349

#> Retrieving page 274 of 349

#> Retrieving page 275 of 349

#> Retrieving page 276 of 349

#> Retrieving page 277 of 349

#> Retrieving page 278 of 349

#> Retrieving page 279 of 349

#> Retrieving page 280 of 349

#> Retrieving page 281 of 349

#> Retrieving page 282 of 349

#> Retrieving page 283 of 349

#> Retrieving page 284 of 349

#> Retrieving page 285 of 349

#> Retrieving page 286 of 349

#> Retrieving page 287 of 349

#> Retrieving page 288 of 349

#> Retrieving page 289 of 349

#> Retrieving page 290 of 349

#> Retrieving page 291 of 349

#> Retrieving page 292 of 349

#> Retrieving page 293 of 349

#> Retrieving page 294 of 349

#> Retrieving page 295 of 349

#> Retrieving page 296 of 349

#> Retrieving page 297 of 349

#> Retrieving page 298 of 349

#> Retrieving page 299 of 349

#> Retrieving page 300 of 349

#> Retrieving page 301 of 349

#> Retrieving page 302 of 349

#> Retrieving page 303 of 349

#> Retrieving page 304 of 349

#> Retrieving page 305 of 349

#> Retrieving page 306 of 349

#> Retrieving page 307 of 349

#> Retrieving page 308 of 349

#> Retrieving page 309 of 349

#> Retrieving page 310 of 349

#> Retrieving page 311 of 349

#> Retrieving page 312 of 349

#> Retrieving page 313 of 349

#> Retrieving page 314 of 349

#> Retrieving page 315 of 349

#> Retrieving page 316 of 349

#> Retrieving page 317 of 349

#> Retrieving page 318 of 349

#> Retrieving page 319 of 349

#> Retrieving page 320 of 349

#> Retrieving page 321 of 349

#> Retrieving page 322 of 349

#> Retrieving page 323 of 349

#> Retrieving page 324 of 349

#> Retrieving page 325 of 349

#> Retrieving page 326 of 349

#> Retrieving page 327 of 349

#> Retrieving page 328 of 349

#> Retrieving page 329 of 349

#> Retrieving page 330 of 349

#> Retrieving page 331 of 349

#> Retrieving page 332 of 349

#> Retrieving page 333 of 349

#> Retrieving page 334 of 349

#> Retrieving page 335 of 349

#> Retrieving page 336 of 349

#> Retrieving page 337 of 349

#> Retrieving page 338 of 349

#> Retrieving page 339 of 349

#> Retrieving page 340 of 349

#> Retrieving page 341 of 349

#> Retrieving page 342 of 349

#> Retrieving page 343 of 349

#> Retrieving page 344 of 349

#> Retrieving page 345 of 349

#> Retrieving page 346 of 349

#> Retrieving page 347 of 349

#> Retrieving page 348 of 349

#> Retrieving page 349 of 349